perf: webgpu grow-only pooled staging/GPU buffers + mock-device integration tests; refactor: split svg backend into 8 modules (exports unchanged)
